% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnaSeqTutorials-package.R
\docType{package}
\name{RnaSeqTutorial01 package}
\alias{RnaSeqTutorial01}
\alias{RnaSeqTutorial01-package}
\alias{RnaSeqTutorial01 package}
\alias{RnaSeqTutorials-package}
\title{RnaSeqTutorials package}
\description{
A simple description of the RnaSeqTutorials package
}
\section{Tutorials}{

This is the first in a series of tutorials
\itemize{
\item\code{01_data_science_intro} an introduction to data science using the palmerpenguins package
}
}

\examples{
	\dontrun{
	learnr::run_tutorial("01_data_science_intro", package = "RnaSeqTutorials")
	}
}
\seealso{
The vignette
}
\author{
Nicolas Delhomme [aut,cre]
}
\keyword{internal}
\keyword{package}
